Migrating to 64-bit. Compiled ok in vs2010, but HANDLE value is still displayed as 32-bit.
git-svn-id: svn://10.22.3.84/windrv-learning/trunk@350 34977f82-1c41-1c46-8be5-c6b895e38539

+++ WinSpy++64/InjectThread.c
@@ -44,8 +44,8 @@ DWORD InjectRemoteThread(HWND hwnd, LPTHREAD_START_ROUTINE lpCode, DWORD cbCodeS
 	HANDLE hRemoteThread = 0;	//handle to the injected thread
 	DWORD  dwRemoteThreadId =0;	//ID of the injected thread
 	
-	DWORD dwWritten = 0;		// Number of bytes written to the remote process
-	DWORD dwRead = 0;
+	SIZE_T dwWritten = 0;		// Number of bytes written to the remote process
+	SIZE_T dwRead = 0;
 	DWORD dwExitCode;
 	
 	void *pRemoteData = 0;
